Use const char ** instead of SEXP for [xyz]names.
The calls to mkChar needed to create a CHARSXPs from pointers to const
char when creating the columns of the resulting data frame are cheap
because mkChar uses caching.  The names that we want to convert to
CHARSXP are the column names of the x, y, and z matrices and therefore
already exist as CHARSXPs.  Thus every call to mkChar will simply return
a cached CHARSXP.